PaxosProducts/CellConfig/src/cellcfg/requirements.txt
blinker==1.4
cmdbpeewee==0.2.1
Fabric3==1.11.1.post1
Flask==0.11.1
future==0.15.2
gevent==1.1.1
peewee==2.8.0
paramiko==2.0.1
pycmdb==0.2.2
pycrypto==2.6.1
pycss==0.2.4
WTForms==2.1
WTForms-ParsleyJS==2.0.1
